investmentapp/requirements.txt
alpaca-py==0.16.0
annotated-types==0.6.0
apsw==3.45.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==2.1.1
finsymbols==1.3.0
idna==3.6
lxml==5.3.1
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sseclient-py==1.8.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==11.0.3
